cne_docker/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pandas==2.2.2
python-docx==1.1.2
rapidfuzz==3.9.7
pydantic==2.8.2
